update build process (#82)
* initial draft of documentation

This commit adds the initial draft of the documentation, including a dedicated docs directory that stores all documentation-related files (content and setup), a .gitignore that excludes documentation build files from tracking and a workflow file that automatically builds and publishes the documentation when the main branch is updated.

* update build process

This commit addresses an error concerning the build process, specifically, we need a dedicated environment to build the documentation. Therefore, a requirements-dev.txt with all dependencies was added and the workflow adapted to install them before attempting to build the documentation.